Add initial lookup logic for results
With the initial logic in place, we should be able to start
getting the raw data for each team from 2010 and on. Currently the end year is set to 2011 for testing purposes
until I can get a new API key. This should allow
us to start getting the data before merging anything and getting our actual overlaps visualized